# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Agriculture API Docs",
        "description": "API Docs for Agriculture Reports",
        "contact": {
            "email": "sajeeshe@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/reports/list": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "List All Reports",
                "responses": {
                    "200": {
                        "description": "Return Reports list"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/report/add": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Add Report Entry",
                "parameters": [
                    {
                        "name": "tractor_name",
                        "in": "query",
                        "description": "Tractor name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "report_date",
                        "in": "query",
                        "description": "Report Date",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "field_name",
                        "in": "query",
                        "description": "Field Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "processed_area",
                        "in": "query",
                        "description": "Processed Area",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Created Field"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/report/show/{slug}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Show selected Report",
                "responses": {
                    "200": {
                        "description": "Return Selected Report"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/report/update/{slug}": {
            "put": {
                "tags": [
                    "Reports"
                ],
                "summary": "Update Report Entry",
                "parameters": [
                    {
                        "name": "tractor_name",
                        "in": "query",
                        "description": "Tractor name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "report_date",
                        "in": "query",
                        "description": "Report Date",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "field_name",
                        "in": "query",
                        "description": "Field Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "processed_area",
                        "in": "query",
                        "description": "Processed Area",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Report Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Created Field"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/report/approve/{slug}": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Approve a Report ( Supervisor | Admin )",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Report Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return success message"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/report/delete/{slug}": {
            "delete": {
                "tags": [
                    "Reports"
                ],
                "summary": "Delete a Report ( Supervisor | Admin )",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Report Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return success message"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/reports/view": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "View the Report ",
                "responses": {
                    "200": {
                        "description": "Return the report list"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}